backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning and Data Processing
tensorflow==2.15.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# API and Data Fetching
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Database (optional, for storing predictions/models)
sqlalchemy==2.0.23
databases[sqlite]==0.8.0
alembic==1.13.1

# Async processing and background tasks
celery==5.3.4
redis==5.0.1

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0

# File handling and utilities
python-dateutil==2.8.2
pytz==2023.3
pillow==10.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS and middleware
python-cors==1.7.0